+++ src/acp/control-plane/manager.runtime-config.test.ts
@@ -268,131 +268,6 @@ describe("AcpSessionManager runtime config", () => {
     });
   });
 
-  it("reconciles pending ACP identities during startup scan", async () => {
-    const runtimeState = createRuntime();
-    runtimeState.getStatus.mockResolvedValue({
-      summary: "status=alive",
-      acpxRecordId: "acpx-record-1",
-      backendSessionId: "acpx-session-1",
-      agentSessionId: "agent-session-1",
-      details: { status: "alive" },
-    });
-    hoisted.requireAcpRuntimeBackendMock.mockReturnValue({
-      id: "acpx",
-      runtime: runtimeState.runtime,
-    });
-
-    let currentMeta: SessionAcpMeta = {
-      ...readySessionMeta(),
-      identity: {
-        state: "pending",
-        source: "ensure",
-        acpxSessionId: "acpx-stale",
-        lastUpdatedAt: Date.now(),
-      },
-    };
-    const sessionKey = "agent:codex:acp:session-1";
-    hoisted.listAcpSessionEntriesMock.mockResolvedValue([
-      {
-        cfg: baseCfg,
-        storePath: "/tmp/sessions-acp.json",
-        sessionKey,
-        storeSessionKey: sessionKey,
-        entry: {
-          sessionId: "session-1",
-          updatedAt: Date.now(),
-          acp: currentMeta,
-        },
-        acp: currentMeta,
-      },
-    ]);
-    hoisted.readAcpSessionEntryMock.mockImplementation((paramsUnknown: unknown) => {
-      const key = (paramsUnknown as { sessionKey?: string }).sessionKey ?? sessionKey;
-      return {
-        sessionKey: key,
-        storeSessionKey: key,
-        acp: currentMeta,
-      };
-    });
-    hoisted.upsertAcpSessionMetaMock.mockImplementation(async (paramsUnknown: unknown) => {
-      const params = paramsUnknown as {
-        mutate: (
-          current: SessionAcpMeta | undefined,
-          entry: { acp?: SessionAcpMeta } | undefined,
-        ) => SessionAcpMeta | null | undefined;
-      };
-      const next = params.mutate(currentMeta, { acp: currentMeta });
-      if (next) {
-        currentMeta = next;
-      }
-      return {
-        sessionId: "session-1",
-        updatedAt: Date.now(),
-        acp: currentMeta,
-      };
-    });
-
-    const manager = new AcpSessionManager();
-    const result = await manager.reconcilePendingSessionIdentities({ cfg: baseCfg });
-
-    expect(result).toEqual({ checked: 1, resolved: 1, failed: 0 });
-    expect(currentMeta.identity?.state).toBe("resolved");
-    expect(currentMeta.identity?.acpxRecordId).toBe("acpx-record-1");
-    expect(currentMeta.identity?.acpxSessionId).toBe("acpx-session-1");
-    expect(currentMeta.identity?.agentSessionId).toBe("agent-session-1");
-  });
-
-  it("skips startup reconcile for pending identities without stable runtime ids", async () => {
-    const runtimeState = createRuntime();
-    hoisted.requireAcpRuntimeBackendMock.mockReturnValue({
-      id: "acpx",
-      runtime: runtimeState.runtime,
-    });
-
-    const sessionKey = "agent:claude:acp:binding:discord:default:9373ab192b2317f4";
-    hoisted.listAcpSessionEntriesMock.mockResolvedValue([
-      {
-        cfg: baseCfg,
-        storePath: "/tmp/sessions-acp.json",
-        sessionKey,
-        storeSessionKey: sessionKey,
-        entry: {
-          sessionId: "session-1",
-          updatedAt: Date.now(),
-          acp: {
-            ...readySessionMeta({
-              agent: "claude",
-            }),
-            identity: {
-              state: "pending",
-              acpxRecordId: sessionKey,
-              source: "status",
-              lastUpdatedAt: Date.now(),
-            },
-          },
-        },
-        acp: {
-          ...readySessionMeta({
-            agent: "claude",
-          }),
-          identity: {
-            state: "pending",
-            acpxRecordId: sessionKey,
-            source: "status",
-            lastUpdatedAt: Date.now(),
-          },
-        },
-      },
-    ]);
-
-    const manager = new AcpSessionManager();
-    const result = await manager.reconcilePendingSessionIdentities({ cfg: baseCfg });
-
-    expect(result).toEqual({ checked: 0, resolved: 0, failed: 0 });
-    expect(runtimeState.ensureSession).not.toHaveBeenCalled();
-    expect(runtimeState.getStatus).not.toHaveBeenCalled();
-  });
-
   it("reconciles prompt-learned agent session IDs even when runtime status omits them", async () => {
     const runtimeState = createRuntime();
     runtimeState.ensureSession.mockResolvedValue({
@@ -468,46 +343,6 @@ describe("AcpSessionManager runtime config", () => {
     expect(currentMeta.identity?.state).toBe("resolved");
     expect(currentMeta.identity?.agentSessionId).toBe("gemini-session-1");
     expect(currentMeta.identity?.acpxSessionId).toBe("acpx-stale");
-  });
-
-  it("skips startup identity reconciliation for already resolved sessions", async () => {
-    const runtimeState = createRuntime();
-    hoisted.requireAcpRuntimeBackendMock.mockReturnValue({
-      id: "acpx",
-      runtime: runtimeState.runtime,
-    });
-    const sessionKey = "agent:codex:acp:session-1";
-    const resolvedMeta: SessionAcpMeta = {
-      ...readySessionMeta(),
-      identity: {
-        state: "resolved",
-        source: "status",
-        acpxSessionId: "acpx-sid-1",
-        agentSessionId: "agent-sid-1",
-        lastUpdatedAt: Date.now(),
-      },
-    };
-    hoisted.listAcpSessionEntriesMock.mockResolvedValue([
-      {
-        cfg: baseCfg,
-        storePath: "/tmp/sessions-acp.json",
-        sessionKey,
-        storeSessionKey: sessionKey,
-        entry: {
-          sessionId: "session-1",
-          updatedAt: Date.now(),
-          acp: resolvedMeta,
-        },
-        acp: resolvedMeta,
-      },
-    ]);
-
-    const manager = new AcpSessionManager();
-    const result = await manager.reconcilePendingSessionIdentities({ cfg: baseCfg });
-
-    expect(result).toEqual({ checked: 0, resolved: 0, failed: 0 });
-    expect(runtimeState.getStatus).not.toHaveBeenCalled();
-    expect(runtimeState.ensureSession).not.toHaveBeenCalled();
   });
 
   it("preserves existing ACP session identifiers when ensure returns none", async () => {
