+++ src/acp/control-plane/manager.status.ts
@@ -1,0 +1,89 @@
+import { resolveSessionIdentityFromMeta } from "@openclaw/acp-core/runtime/session-identity";
+import type {
+  AcpRuntime,
+  AcpRuntimeCapabilities,
+  AcpRuntimeHandle,
+  AcpRuntimeStatus,
+} from "@openclaw/acp-core/runtime/types";
+import type { OpenClawConfig } from "../../config/types.openclaw.js";
+import { withAcpRuntimeErrorBoundary } from "../runtime/errors.js";
+import type {
+  AcpSessionStatus,
+  EnsureManagerRuntimeHandle,
+  ReconcileManagerRuntimeSessionIdentifiers,
+  ResolveManagerSession,
+} from "./manager.types.js";
+import { requireReadySessionMeta } from "./manager.utils.js";
+import { resolveRuntimeOptionsFromMeta } from "./runtime-options.js";
+
+export async function runManagerGetSessionStatus(params: {
+  cfg: OpenClawConfig;
+  sessionKey: string;
+  signal?: AbortSignal;
+  throwIfAborted: (signal?: AbortSignal) => void;
+  resolveSession: ResolveManagerSession;
+  ensureRuntimeHandle: EnsureManagerRuntimeHandle;
+  resolveRuntimeCapabilities: (params: {
+    runtime: AcpRuntime;
+    handle: AcpRuntimeHandle;
+  }) => Promise<AcpRuntimeCapabilities>;
+  reconcileRuntimeSessionIdentifiers: ReconcileManagerRuntimeSessionIdentifiers;
+}): Promise<AcpSessionStatus> {
+  params.throwIfAborted(params.signal);
+  const resolution = params.resolveSession({
+    cfg: params.cfg,
+    sessionKey: params.sessionKey,
+  });
+  const resolvedMeta = requireReadySessionMeta(resolution);
+  const {
+    runtime,
+    handle: ensuredHandle,
+    meta: ensuredMeta,
+  } = await params.ensureRuntimeHandle({
+    cfg: params.cfg,
+    sessionKey: params.sessionKey,
+    meta: resolvedMeta,
+  });
+  let handle = ensuredHandle;
+  let meta = ensuredMeta;
+  const capabilities = await params.resolveRuntimeCapabilities({ runtime, handle });
+  let runtimeStatus: AcpRuntimeStatus | undefined;
+  if (runtime.getStatus) {
+    runtimeStatus = await withAcpRuntimeErrorBoundary({
+      run: async () => {
+        params.throwIfAborted(params.signal);
+        const status = await runtime.getStatus!({
+          handle,
+          ...(params.signal ? { signal: params.signal } : {}),
+        });
+        params.throwIfAborted(params.signal);
+        return status;
+      },
+      fallbackCode: "ACP_TURN_FAILED",
+      fallbackMessage: "Could not read ACP runtime status.",
+    });
+  }
+  ({ handle, meta, runtimeStatus } = await params.reconcileRuntimeSessionIdentifiers({
+    cfg: params.cfg,
+    sessionKey: params.sessionKey,
+    runtime,
+    handle,
+    meta,
+    runtimeStatus,
+    failOnStatusError: true,
+  }));
+  const identity = resolveSessionIdentityFromMeta(meta);
+  return {
+    sessionKey: params.sessionKey,
+    backend: handle.backend || meta.backend,
+    agent: meta.agent,
+    ...(identity ? { identity } : {}),
+    state: meta.state,
+    mode: meta.mode,
+    runtimeOptions: resolveRuntimeOptionsFromMeta(meta),
+    capabilities,
+    runtimeStatus,
+    lastActivityAt: meta.lastActivityAt,
+    lastError: meta.lastError,
+  };
+}
