+++ src/acp/control-plane/manager.types.ts
@@ -159,6 +159,41 @@ export type WriteManagerSessionMeta = (params: {
   takeCacheOwnership?: boolean;
 }) => Promise<SessionEntry | null>;
 
+export type ResolveManagerSession = (params: {
+  cfg: OpenClawConfig;
+  sessionKey: string;
+}) => AcpSessionResolution;
+
+export type EnsureManagerRuntimeHandle = (params: {
+  cfg: OpenClawConfig;
+  sessionKey: string;
+  meta: SessionAcpMeta;
+}) => Promise<{ runtime: AcpRuntime; handle: AcpRuntimeHandle; meta: SessionAcpMeta }>;
+
+export type ReconcileManagerRuntimeSessionIdentifiers = (params: {
+  cfg: OpenClawConfig;
+  sessionKey: string;
+  runtime: AcpRuntime;
+  handle: AcpRuntimeHandle;
+  meta: SessionAcpMeta;
+  runtimeStatus?: AcpRuntimeStatus;
+  failOnStatusError: boolean;
+}) => Promise<{
+  handle: AcpRuntimeHandle;
+  meta: SessionAcpMeta;
+  runtimeStatus?: AcpRuntimeStatus;
+}>;
+
+export type SetManagerSessionState = (params: {
+  cfg: OpenClawConfig;
+  sessionKey: string;
+  state: SessionAcpMeta["state"];
+  lastError?: string;
+  clearLastError?: boolean;
+}) => Promise<void>;
+
+export type WithManagerSessionActor = <T>(sessionKey: string, op: () => Promise<T>) => Promise<T>;
+
 export const DEFAULT_DEPS: AcpSessionManagerDeps = {
   listAcpSessions: listAcpSessionEntries,
   readSessionEntry: readAcpSessionEntry,
